+++ package.json
@@ -1,10 +1,12 @@
 {
-  "name": "cart-app",
+  "name": "shopping-web",
   "version": "1.0.0",
   "description": "",
   "main": "server.ts",
   "scripts": {
-    "dev": "nodemon --exec ts-node src/server.ts"
+    "dev": "nodemon --exec ts-node src/server.ts",
+    "build": "tsc",
+    "start": "node dist/server.js"
   },
   "keywords": [],
   "author": "",
