+++ src/admin/controller/roleController.ts
@@ -1,18 +1,40 @@
 import { Request, Response } from "express";
-import * as roleServices from "../services/roleService";
-import { StatusCodes, ErrorMessages } from "../../validation/responseMessages";
+import {
+  ErrorMessages,
+  StatusCodes,
+  SuccessMessages,
+} from "../../validation/responseMessages";
+import { Role } from "../models/roleModel";
 
-// Admin Register
+// create Role
 export const createRole = async (req: Request, res: Response) => {
   try {
-    const result = await roleServices.createRole(req.body);
-    return res.status(result.status).json(result);
+    const { role } = req.body;
+    const existingRole = await Role.findOne({ role });
+    if (existingRole) {
+      return res.json({
+        message: ErrorMessages.RoleExist,
+        success: false,
+        status: StatusCodes.ClientError.BadRequest,
+      });
+    }
+
+    const newRole = new Role({
+      role,
+    });
+
+    await newRole.save();
+    return res.json({
+      message: SuccessMessages.RoleCreated,
+      status: StatusCodes.Success.Created,
+      success: true,
+    });
   } catch (error) {
-    console.error("Error in create role", error);
-    return {
+    console.error("Error in register", error);
+    return res.json({
       message: ErrorMessages.SomethingWentWrong,
       success: false,
       status: StatusCodes.ServerError.InternalServerError,
-    };
+    });
   }
 };
