+++ src/helpers/logger.ts
@@ -28,11 +28,11 @@ const format = winston.format.combine(
 
 const transports = [
   new winston.transports.Console(),
-  new winston.transports.File({
-    filename: "logs/error.log",
-    level: "error",
-  }),
-  new winston.transports.File({ filename: "logs/all.log" }),
+  // new winston.transports.File({
+  //   filename: "logs/error.log",
+  //   level: "error",
+  // }),
+  // new winston.transports.File({ filename: "logs/all.log" }),
 ];
 
 const Logger = winston.createLogger({
